Preserves selection state on refresh (when possible)
If the WorkingSetView is refreshed, and the search term has not changed,
then WS will now check to see if the previously selected item is in the
new search results, and if so, it'll set it so that item is selected and
in view.

This will make live-update mode more useful, as the view will not reset
everytime a file is saved.