Adds handling of window resize
On KEY_RESIZE, the current view is re-rendered.  This improves the
experience especially when the window size is increased, revealing
portions of the window that are blank, but wouldn't be after a render at
the new window size.

Also refactored some of the generic view switching to a view-router
actor.